TST: Rearrange discovery tests into submodules.
